Rename real_time_graph/ metric/.
